API PUBLISH

+++ MyWebAPI/Model/P_MarketPrice.cs
@@ -7,6 +7,6 @@
         public string Location { get; set; }
 
         public int Capacity { get; set; }
-        public int Price_100g { get; set; }
+        public double Price_100g { get; set; }
     }
 }
